+++ api/internal/auth/repository/repository.go
@@ -1,12 +1,6 @@
 package repository
 
-import (
-	"context"
-	"database/sql"
-	"errors"
-
-	"github.com/Baymax6s/KOBE-Tech/api/internal/auth"
-)
+import "database/sql"
 
 type Repository struct {
 	db *sql.DB
@@ -14,56 +8,4 @@ type Repository struct {
 
 func NewRepository(db *sql.DB) *Repository {
 	return &Repository{db: db}
-}
-
-func (r *Repository) FindByID(ctx context.Context, id int64) (auth.User, error) {
-	if r == nil || r.db == nil {
-		return auth.User{}, errors.New("auth repository is not configured")
-	}
-
-	const query = `
-		SELECT id, name, password_hash, created_at, updated_at
-		FROM users
-		WHERE id = $1
-	`
-
-	var user auth.User
-	err := r.db.QueryRowContext(ctx, query, id).Scan(
-		&user.ID,
-		&user.Name,
-		&user.PasswordHash,
-		&user.CreatedAt,
-		&user.UpdatedAt,
-	)
-	if err != nil {
-		return auth.User{}, err
-	}
-
-	return user, nil
-}
-
-func (r *Repository) FindByName(ctx context.Context, name string) (auth.User, error) {
-	if r == nil || r.db == nil {
-		return auth.User{}, errors.New("auth repository is not configured")
-	}
-
-	const query = `
-		SELECT id, name, password_hash, created_at, updated_at
-		FROM users
-		WHERE name = $1
-	`
-
-	var user auth.User
-	err := r.db.QueryRowContext(ctx, query, name).Scan(
-		&user.ID,
-		&user.Name,
-		&user.PasswordHash,
-		&user.CreatedAt,
-		&user.UpdatedAt,
-	)
-	if err != nil {
-		return auth.User{}, err
-	}
-
-	return user, nil
 }
